+++ src/main/java/com/api/mysql/api_try3/repository/EventRepository.java
@@ -8,9 +8,9 @@ import org.springframework.data.repository.query.Param;
 import java.sql.Date;
 import java.util.List;
 
-public interface EventRepository extends JpaRepository<Event, Long> {
+public interface EventRepository extends JpaRepository<Event, Integer> {
     @Query("SELECT e FROM Event e WHERE e.idRoom = :idRoom AND e.date = :date AND e.time = :time AND e.place = :place AND e.event = :event")
-    Event findEventByCriteria(@Param("idRoom") Long idRoom, @Param("date") String date, @Param("time") String time, @Param("place") String place, @Param("event") String event);
+    Event findEventByCriteria(@Param("idRoom") int idRoom, @Param("date") String date, @Param("time") String time, @Param("place") String place, @Param("event") String event);
 
-    List<Event> findByIdRoom(Long idRoom);
+    List<Event> findByIdRoom(int idRoom);
 }
