+++ src/main/java/com/api/mysql/api_try3/repository/FileRepository.java
@@ -7,9 +7,9 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface FileRepository extends JpaRepository<File, Long> {
+public interface FileRepository extends JpaRepository<File, Integer> {
     @Query("SELECT i FROM File i WHERE i.idRoom = :idRoom AND i.date = :date AND i.time = :time AND i.url = :url")
-    File findFileByCriteria(@Param("idRoom") Long idRoom, @Param("date") String date, @Param("time") String time, @Param("url") String url);
+    File findFileByCriteria(@Param("idRoom") int idRoom, @Param("date") String date, @Param("time") String time, @Param("url") String url);
 
-    List<File> findByIdRoom(Long idRoom);
+    List<File> findByIdRoom(int idRoom);
 }
