+++ src/main/java/com/api/mysql/api_try3/repository/TaskRepository.java
@@ -8,9 +8,9 @@ import org.springframework.data.repository.query.Param;
 import java.sql.Date;
 import java.util.List;
 
-public interface TaskRepository extends JpaRepository<Task, Long> {
+public interface TaskRepository extends JpaRepository<Task, Integer> {
     @Query("SELECT t FROM Task t WHERE t.idRoom = :idRoom AND t.date = :date AND t.time = :time AND t.name = :name AND t.points = :points AND t.checkBoxes = :checkBoxes")
-    List<Task> findTaskByCriteria(@Param("idRoom") Long idRoom, @Param("date") String date, @Param("time") String time, @Param("name") String name, @Param("points") String points, @Param("checkBoxes") String checkBoxes);
+    List<Task> findTaskByCriteria(@Param("idRoom") int idRoom, @Param("date") String date, @Param("time") String time, @Param("name") String name, @Param("points") String points, @Param("checkBoxes") String checkBoxes);
 
-    List<Task> findByIdRoom(Long idRoom);
+    List<Task> findByIdRoom(int idRoom);
 }
