+++ src/main/java/com/api/mysql/api_try3/service/imlp/ImageServiceImpl.java
@@ -20,14 +20,14 @@ public class ImageServiceImpl {
         imageRepository.save(image);
     }
 
-    public void deleteImage(Long idImage) {
+    public void deleteImage(int idImage) {
         imageRepository.deleteById(idImage);
     }
-    public Image getImage(Long idRoom, String date, String time, String url) {
+    public Image getImage(int idRoom, String date, String time, String url) {
         return imageRepository.findImageByCriteria(idRoom, date, time, url);
     }
 
-    public List<Image> getImagesByIdRoom(Long idRoom) {
+    public List<Image> getImagesByIdRoom(int idRoom) {
         return imageRepository.findByIdRoom(idRoom);
     }
